+++ packages/cubejs-api-gateway/src/helpers/prepareAnnotation.ts
@@ -48,7 +48,8 @@ const annotation = (
 ) => (member: string | MemberExpression): undefined | [string, ConfigItem] => {
   const [cubeName, fieldName] = (<MemberExpression>member).expression ? [(<MemberExpression>member).cubeName, (<MemberExpression>member).name] : (<string>member).split('.');
   const memberWithoutGranularity = [cubeName, fieldName].join('.');
-  const config: ConfigItem = configMap[cubeName][memberType]
+  const cubeConfig = configMap[cubeName];
+  const config: ConfigItem = cubeConfig && cubeConfig[memberType]
     .find(m => m.name === memberWithoutGranularity);
 
   if (!config) {