+++ packages/cubejs-playground/src/QueryBuilderV2/hooks/server-core-version-gte.ts
@@ -3,10 +3,13 @@ export function useServerCoreVersionGte(version: string, currentVersion: string)
     let gt = false;
 
     try {
-      const [, m, p] = currentVersion.split('.').map(Number);
-      const [, m1, p1] = version.split('.').map(Number);
+      const [major, minor, patch] = currentVersion.split('.').map(Number);
+      const [major1, minor1, patch1] = version.split('.').map(Number);
 
-      gt = m > m1 || (m === m1 && p >= p1);
+      gt =
+        major > major1 ||
+        (major === major1 && minor > minor1) ||
+        (major === major1 && minor === minor1 && patch >= patch1);
     } catch (_) {
       //
     }