+++ packages/cubejs-schema-compiler/test/unit/yaml-schema.test.ts
@@ -348,4 +348,53 @@ describe('Yaml Schema Testing', () => {
       await compiler.compile();
     });
   });
+  describe('Access policy: ', () => {
+    it('defines a correct accessPolicy', async () => {
+      const { compiler } = prepareYamlCompiler(
+        `
+        cubes:
+        - name: Orders
+          sql: "select * from tbl"
+          dimensions:
+            - name: created_at
+              sql: created_at
+              type: time
+            - name: status
+              sql: status
+              type: string
+          measures:
+            - name: count
+              type: count
+          accessPolicy:
+            - role: admin
+              conditions:
+                - if: "{ !security_context.isBlocked }"
+              rowLevel:
+                filters:
+                  - member: status
+                    operator: equals
+                    values: ["completed"]
+                  - or:
+                    - member: "{CUBE}.created_at"
+                      operator: notInDateRange
+                      values:
+                        - 2022-01-01
+                        - "{ security_context.currentDate }"
+                    - member: "created_at"
+                      operator: equals
+                      values:
+                        - "{ securityContext.currentDate }"
+              memberLevel:
+                includes:
+                  - status
+            - role: manager
+              memberLevel:
+                excludes:
+                  - status
+        `
+      );
+
+      await compiler.compile();
+    });
+  });
 });